requirements.txt
# Core training & inference stack
torch>=2.3.0
transformers>=4.39.0
modelscope>=1.13.0
vllm>=0.5.4
segment-anything>=1.0

# Vision & document processing
opencv-python-headless>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
pdf2image>=1.17.0
PyMuPDF>=1.24.0  # provides fitz for fast PDF rasterization
tqdm>=4.65.0
matplotlib>=3.8.0

# Data & model utilities
requests>=2.31.0
feedparser>=6.0.10
qwen-vl-utils>=0.0.8
lmdeploy>=0.5.0
openai>=1.23.0
